Fixed Update Issue
Issue description:
Updating any restaurant entity in Angular caused the loss of all rows in the table that connected restaurants with the products they offered (drinks menus and food menus). -Fixed by instantly instantiating the sets of data that containt that information (Set<ProductBean> foodMenu/drinksMenu in the RestaurantBean class)
Updating any restaurant entity THAT WAS RELATED TO A MENU cause the loss of all data in the merginig table (same as above). After the inital fix, the only problem were the tables that were connected to any products. -Fixed by specifying what data should be overwriten in the controller (Controller took the entire object that was passed in angluar, this meant that the overwritten object had no information of products. Took object from database by key, changed name and type, wrote back into the database.)