Restaurant Editing[0]
Added a Restaurant type Table that contains possible restaurant types. A restaurant can be more than one type (?), so I had to add this and connect it to the restaurant via relation. Now a checkbox or pick list can be implemented in editing a restaurant (Choosing which type(s) a restaurant is). This can also be implemented for which types of food a restaurant serves, or not, if we're going by menu.