Restaurant Editing[4]
Succesfully implemented basic ng components for chaning the restaurant name (a text input), and the type of the restaurant (a select button group -> only one can be active). Added toolbar for saving or discarding changes, implement closing the panel after any of these buttons are clicked, with the respected function of either saving or discarding changes.