Restaurant Editing[9]
Completed basic editing of a restaurant. Food and beverage items can be added and removed from a restaurant. The restaurants name can be changed, and the type of the restaurant can be edited. All other information that can be changed about a restaurant needs only to be added (All things are researched, other components and functionalities can be added following the current functionality implementation)