Restaurant Editing[10]
Changed view of the restaurant Editing component. Restaurant editing begins on edit button in data grid. The grid is hidden and the panel for editing appears with tabs on sections about the restaurant to edit.
NOTE(1) : All commits related to the view component of restaurant editing will be marked with Restaurant View[X], and will not have a commit body. Only functionalities will have commit comments and will proceed to be marked as Restaurant Editing[X]