Restaurant View[0]
Began adding normal images to restaurants (not from lorem-fing-pixel). Restaurants image will be saved as a string representing the file path to the image. Restaurants that do not have a profile picture will be given 'null' for the file path. Database holds the file path to the iamge. The image must begin with /assets/pictures/restaurant_pictures/ as it is the directory where all restaurant pictures will be saved. The name of the image must be the code of the restaurant. Restaurants that will have their image edited will not have ANOTHER image in the database and folder but will save the new image OVER the existing one. Restaurants that do ot have a valid image (null) will read the default /assets/pictures/no_image_found.jpg picture.