Image Upload for Restaurant
Finally uploaded a goddamn image to the back end. Problem was no header in the RESPONSE - or rather - an incorrect header. Request headers content-type set to multipart/* in request mapping, and inserted 'artifical' header in the response headers - the information of the file and the file name.
Sloppy change of the image file name, but it works. Will work on changing to a more ellegant way in the future. All database entries must have an image path, if the image does not exist it wont be shown anyway, and database update is no longer needed for changing image.
Sloppy coding in angular, will change next time.
Changed look of restaurant browsing UI - Now showing restaurants cuisines and it's a generally better look and feel. Image related UI will be updated next time.
NOTE - GENERIC PATH NOT SET. Uploading images will always return an error. Just don't upload anything now and i'll set it next time.
The Emperor Protects.'